optimize for my self

+++ app/src/main/java/com/example/villageevo/domain/building/BuildingDefinition.kt
@@ -1,15 +1,12 @@
 package com.example.villageevo.domain.building
 
-import com.example.villageevo.domain.worker.WorkerRole
-
 data class BuildingDefinition(
-    val type: BuildingType,
-    val name: String,
-    val buildCostGold: Int,
-    val buildCostLumber: Int,
-    val baseProduction: Int,
-    val baseOperationalCost: Int,
-    val requiredWorkers: Int,
-    val description: String
-
-    )
+        val type: BuildingType,
+        val name: String,
+        val buildCostGold: Long,
+        val buildCostLumber: Long,
+        val baseProduction: Long,
+        val baseOperationalCost: Long,
+        val requiredWorkers: Int,
+        val description: String
+)
